+++ runtime/js/40_files.js
@@ -159,105 +159,6 @@ class FsFile {
   }
 }
 
-class Stdin {
-  #readable;
-
-  constructor() {
-  }
-
-  get rid() {
-    return 0;
-  }
-
-  read(p) {
-    return read(this.rid, p);
-  }
-
-  readSync(p) {
-    return readSync(this.rid, p);
-  }
-
-  close() {
-    core.close(this.rid);
-  }
-
-  get readable() {
-    if (this.#readable === undefined) {
-      this.#readable = readableStreamForRid(this.rid);
-    }
-    return this.#readable;
-  }
-
-  setRaw(mode, options = {}) {
-    const cbreak = !!(options.cbreak ?? false);
-    ops.op_stdin_set_raw(mode, cbreak);
-  }
-}
-
-class Stdout {
-  #writable;
-
-  constructor() {
-  }
-
-  get rid() {
-    return 1;
-  }
-
-  write(p) {
-    return write(this.rid, p);
-  }
-
-  writeSync(p) {
-    return writeSync(this.rid, p);
-  }
-
-  close() {
-    core.close(this.rid);
-  }
-
-  get writable() {
-    if (this.#writable === undefined) {
-      this.#writable = writableStreamForRid(this.rid);
-    }
-    return this.#writable;
-  }
-}
-
-class Stderr {
-  #writable;
-
-  constructor() {
-  }
-
-  get rid() {
-    return 2;
-  }
-
-  write(p) {
-    return write(this.rid, p);
-  }
-
-  writeSync(p) {
-    return writeSync(this.rid, p);
-  }
-
-  close() {
-    core.close(this.rid);
-  }
-
-  get writable() {
-    if (this.#writable === undefined) {
-      this.#writable = writableStreamForRid(this.rid);
-    }
-    return this.#writable;
-  }
-}
-
-const stdin = new Stdin();
-const stdout = new Stdout();
-const stderr = new Stderr();
-
 function checkOpenOptions(options) {
   if (
     ArrayPrototypeFilter(
@@ -284,16 +185,4 @@ function checkOpenOptions(options) {
 }
 
 const File = FsFile;
-export {
-  create,
-  createSync,
-  File,
-  FsFile,
-  open,
-  openSync,
-  seek,
-  seekSync,
-  stderr,
-  stdin,
-  stdout,
-};
+export { create, createSync, File, FsFile, open, openSync, seek, seekSync };
