+++ runtime/js/90_deno_ns.js
@@ -108,9 +108,9 @@ const denoNs = {
   openSync: files.openSync,
   create: files.create,
   createSync: files.createSync,
-  stdin: files.stdin,
-  stdout: files.stdout,
-  stderr: files.stderr,
+  stdin: io.stdin,
+  stdout: io.stdout,
+  stderr: io.stderr,
   seek: files.seek,
   seekSync: files.seekSync,
   connect: net.connect,
